src/main/webapp/app/entities/voice-call/voice-call.service.ts
import axios from 'axios';

import { IVoiceCall } from '@/shared/model/voice-call.model';

const baseApiUrl = 'api/voice-calls';

export default class VoiceCallService {
  public find(id): Promise<IVoiceCall> {
    return new Promise<IVoiceCall>(resolve => {
      axios.get(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public retrieve(paginationQuery?: any): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(baseApiUrl).then(function(res) {
        resolve(res);
      });
    });
  }

  public delete(id): Promise<any> {
    return new Promise<any>(resolve => {
      axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res);
      });
    });
  }

  public create(entity): Promise<IVoiceCall> {
    return new Promise<IVoiceCall>(resolve => {
      axios.post(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public update(entity): Promise<IVoiceCall> {
    return new Promise<IVoiceCall>(resolve => {
      axios.put(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }
}
